requirements.txt
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
websocket-client>=1.6.0
openai>=1.30.0
pyperclip>=1.8.2
scipy>=1.11.0
python-dotenv>=1.0.0
